Add thumbv7em-freertos-eabihf target to Rust compiler (#18)
Add thumbv7em-freertos-eabihf target to Rust compiler, which now detects freertos from the target to enable freertos dependency.
This eliminates the need for freertos dependency to be hardcoded, or for restricted_std to be declared in the application.
The tighter integration of freertos features means that the code is subject to 'warnings as errors'. Many steps taken to suppress or ignore warnings (those in bindgen generated code are unavoidable).